+++ src/routes/bot/localisation.js
@@ -1,51 +1,4 @@
 var fs = require('fs');
-const {safeErrorLog, getUserName} = require("./utils");
-const {takeRequestKeyboard} = require("../keyboard");
-const mongoose = require("mongoose");
-const Request = mongoose.model('Request');
-
-async function changeRequestLanguage(request, newLanguage, bot) {
-    let fromLanguageChat = getLanguageTGChat(request.language)
-    let toLanguageChat = getLanguageTGChat(newLanguage)
-    if (fromLanguageChat === toLanguageChat) return;
-    let moderatorMsgId, moderatorActionMsgId;
-    try {
-        moderatorMsgId = await bot.forwardMessage(toLanguageChat, request.requesterTG, request.requesterMsgID);
-    } catch (e) { safeErrorLog(e) }
-    let inline_keyboard = await takeRequestKeyboard(request._id);
-    let options = {
-        reply_to_message_id: moderatorMsgId.message_id,
-        reply_markup: JSON.stringify({
-            inline_keyboard
-        })
-    };
-    try {
-        let initiator = getUserName(moderatorMsgId.forward_from);
-        if (initiator.startsWith("@")) {
-            initiator = initiator.substring(1)
-        }
-        moderatorActionMsgId = await bot.sendMessage(
-            toLanguageChat,
-            '№' + request.requestId + '\nініціатор: ' + initiator + '\n#pending',
-            options,
-        );
-    } catch (e) {safeErrorLog(e)}
-
-    try {
-        await bot.deleteMessage(fromLanguageChat, request.moderatorMsgID)
-        await bot.deleteMessage(fromLanguageChat, request.moderatorActionMsgID)
-    } catch (e) {
-        safeErrorLog(e)
-    }
-    await Request.updateOne(
-        request,
-        {
-            language: newLanguage,
-            moderatorActionMsgID: parseInt(moderatorActionMsgId.message_id),
-            moderatorMsgID: parseInt(moderatorMsgId.message_id),
-        },
-    );
-}
 
 function getLanguageTGChat(language) {
     let moderatorChat = process.env.TGMAINCHAT;
@@ -70,5 +23,4 @@ module.exports = {
             });
     },
     getLanguageTGChat: getLanguageTGChat,
-    changeRequestLanguage: changeRequestLanguage,
 }
