+++ src/routes/bot/query-callbacks.js
@@ -18,7 +18,7 @@ const {
     NoCurrentFakes
 } = require('./contstants')
 const {informRequestersWithComment} = require("./message-handlers");
-const { getText, getLanguageTGChat, changeRequestLanguage} = require('./localisation');
+const { getText, getLanguageTGChat} = require('./localisation');
 const mongoose = require("mongoose");
 require('dotenv').config();
 
@@ -27,6 +27,7 @@ const TelegramUser = mongoose.model('TelegramUser');
 const Data = mongoose.model('Data');
 const Escalation = mongoose.model('Escalation');
 const Comment = mongoose.model('Comment');
+const {takeRequestKeyboard} = require("../keyboard");
 
 const onReqTakeQuery = async (callbackQuery, bot) => {
 
@@ -558,6 +559,50 @@ const onUpdateCommentQuery = async (callbackQuery, bot) => {
         } catch (e) { safeErrorLog(e) } 
     }
 }
+
+async function changeRequestLanguage(request, newLanguage, bot) {
+    let fromLanguageChat = getLanguageTGChat(request.language)
+    let toLanguageChat = getLanguageTGChat(newLanguage)
+    if (fromLanguageChat === toLanguageChat) return;
+    let moderatorMsgId, moderatorActionMsgId;
+    try {
+        moderatorMsgId = await bot.forwardMessage(toLanguageChat, request.requesterTG, request.requesterMsgID);
+    } catch (e) { safeErrorLog(e) }
+    let inline_keyboard = await takeRequestKeyboard(request._id);
+    let options = {
+        reply_to_message_id: moderatorMsgId.message_id,
+        reply_markup: JSON.stringify({
+            inline_keyboard
+        })
+    };
+    try {
+        let initiator = getUserName(moderatorMsgId.forward_from);
+        if (initiator.startsWith("@")) {
+            initiator = initiator.substring(1)
+        }
+        moderatorActionMsgId = await bot.sendMessage(
+            toLanguageChat,
+            '№' + request.requestId + '\nініціатор: ' + initiator + '\n#pending',
+            options,
+        );
+    } catch (e) {safeErrorLog(e)}
+
+    try {
+        await bot.deleteMessage(fromLanguageChat, request.moderatorMsgID)
+        await bot.deleteMessage(fromLanguageChat, request.moderatorActionMsgID)
+    } catch (e) {
+        safeErrorLog(e)
+    }
+    await Request.updateOne(
+        request,
+        {
+            language: newLanguage,
+            moderatorActionMsgID: parseInt(moderatorActionMsgId.message_id),
+            moderatorMsgID: parseInt(moderatorMsgId.message_id),
+        },
+    );
+}
+
 
 module.exports = {
     onFakeStatusQuery,
